Climb Push
Added a very simple way of climbing. Run up to a wall with the correct layer ID to climb.

Known Errors: Player character can stick to the wall indefinietly. May be feature or bug. Could potentially be used as another way to go through the level.
Should probably work on the player looking towards the direction they are moving at